requirements.txt
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.20.0
yfinance>=0.2.0
pyarrow>=14.0.0
pyyaml>=6.0
lxml
html5lib
requests
curl_cffi
torch>=2.0.0
mamba-ssm
causal-conv1d
pykan
iisignature
loguru
scikit-learn
matplotlib
websocket-client
ib_insync
ibapi
